+++ src/Voting/xlvoVotingManager2.php
@@ -63,11 +63,15 @@ class xlvoVotingManager2 {
 	 * @param $pin
 	 */
 	public function __construct($pin) {
-		if (empty($pin)) {
+		// TODO: Hotfix! - Removed this to working enter ping gui again
+		/*if (empty($pin)) {
 			throw new ilException("xlvoVotingManager2: Wrong PIN! - 1");
 		}
 
-		$this->initVoting($pin);
+		$this->initVoting($pin);*/
+		if (!empty($pin)) {
+			$this->initVoting($pin);
+		}
 	}