+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [x]
+- Hotfix: Working enter ping gui again
+
 ## [4.3.0]
 - feature: clustering/categorizing of free text questions